move single order solver retry logic into trait and add try_settle_order
